+++ test/test_automated_analysis.py
@@ -472,3 +472,17 @@ class AutomatedAnalysisTest(unittest.TestCase):
             print("")
             print(" ".join(clcode))
             subprocess.check_call(clcode)
+
+    @attr(complete=True)
+    @attr(cmd_validate=True)
+    @attr(cmd=True)
+    def test_validate_cmd(self):
+        """Run update analysis
+        """
+        with make_workdir():
+            clcode = ["mirtop",
+                      "validate",
+                      "../../data/examples/gff/correct_file.gff"]
+            print("")
+            print(" ".join(clcode))
+            subprocess.check_call(clcode)